requirements.txt
hermes_python==0.*
paho-mqtt==1.*
